int xreca

+++ for.c
@@ -3,11 +3,11 @@
 int main()
 {
 
-    int i, j;
+    long int i, j;
 
     for (i = 0, j = 1000; i < j; i++, j--)
     {
-        printf("i = %d, j = %d\n", i, j);
+        printf("i = %ld, j = %ld\n", i, j);
     }
 
     getchar();
